move fx options into func
