+++ cmd/process-agent/subcommands/check/check_options.go
@@ -3,27 +3,19 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-// Package processchecks implements 'agent processchecks'.
-package processchecks
+package check
 
 import (
-	"github.com/spf13/cobra"
 	"go.uber.org/fx"
 
 	"github.com/DataDog/datadog-go/v5/statsd"
 
-	"github.com/DataDog/datadog-agent/cmd/agent/command"
-	"github.com/DataDog/datadog-agent/cmd/agent/common"
-	processCommand "github.com/DataDog/datadog-agent/cmd/process-agent/command"
-	"github.com/DataDog/datadog-agent/cmd/process-agent/subcommands/check"
 	"github.com/DataDog/datadog-agent/comp/core"
 	ipcfx "github.com/DataDog/datadog-agent/comp/core/ipc/fx"
-	log "github.com/DataDog/datadog-agent/comp/core/log/def"
 	secretsfx "github.com/DataDog/datadog-agent/comp/core/secrets/fx"
 	tagger "github.com/DataDog/datadog-agent/comp/core/tagger/def"
-	dualTaggerfx "github.com/DataDog/datadog-agent/comp/core/tagger/fx-dual"
-	workloadfilterfx "github.com/DataDog/datadog-agent/comp/core/workloadfilter/fx"
-	wlmcatalogcore "github.com/DataDog/datadog-agent/comp/core/workloadmeta/collectors/catalog-core"
+	remoteTaggerfx "github.com/DataDog/datadog-agent/comp/core/tagger/fx-remote"
+	wmcatalogremote "github.com/DataDog/datadog-agent/comp/core/workloadmeta/collectors/catalog-remote"
 	workloadmeta "github.com/DataDog/datadog-agent/comp/core/workloadmeta/def"
 	workloadmetafx "github.com/DataDog/datadog-agent/comp/core/workloadmeta/fx"
 	"github.com/DataDog/datadog-agent/comp/forwarder/eventplatform/eventplatformimpl"
@@ -34,7 +26,7 @@ import (
 	proccontainers "github.com/DataDog/datadog-agent/pkg/process/util/containers"
 )
 
-func getCoreAgentFxOptions(cliParams *check.CliParams, bundleParams core.BundleParams) []fx.Option {
+func getProcessAgentFxOptions(cliParams *CliParams, bundleParams core.BundleParams) []fx.Option {
 	return []fx.Option{
 		fx.Supply(cliParams, bundleParams),
 		core.Bundle(),
@@ -48,15 +40,14 @@ func getCoreAgentFxOptions(cliParams *check.CliParams, bundleParams core.BundleP
 		// Provide npcollector module
 		npcollectorimpl.Module(),
 		// Provide the corresponding workloadmeta Params to configure the catalog
-		wlmcatalogcore.GetCatalog(),
+		wmcatalogremote.GetCatalog(),
 		// Provide workloadmeta module
 		workloadmetafx.Module(workloadmeta.Params{
-			AgentType: workloadmeta.NodeAgent,
+			AgentType: workloadmeta.Remote,
 		}),
-		workloadfilterfx.Module(),
 
 		// Tagger must be initialized after agent config has been setup
-		dualTaggerfx.Module(common.DualTaggerParams()),
+		remoteTaggerfx.Module(tagger.NewRemoteParams()),
 		processComponent.Bundle(),
 		// InitSharedContainerProvider must be called before the application starts so the workloadmeta collector can be initiailized correctly.
 		// Since the tagger depends on the workloadmeta collector, we can not make the tagger a dependency of workloadmeta as it would create a circular dependency.
@@ -70,23 +61,4 @@ func getCoreAgentFxOptions(cliParams *check.CliParams, bundleParams core.BundleP
 		}),
 		ipcfx.ModuleReadOnly(),
 	}
-}
-
-// Commands returns a slice of subcommands for the 'agent' command.
-func Commands(globalParams *command.GlobalParams) []*cobra.Command {
-	processCommand.OneShotLogParams = log.ForOneShot(string(command.LoggerName), "info", true)
-	checkAllowlist := []string{"process", "rtprocess", "container", "rtcontainer", "process_discovery"}
-	cmd := check.MakeCommand(func() *processCommand.GlobalParams {
-		return &processCommand.GlobalParams{
-			ConfFilePath:         globalParams.ConfFilePath,
-			ExtraConfFilePath:    globalParams.ExtraConfFilePath,
-			SysProbeConfFilePath: globalParams.SysProbeConfFilePath,
-			FleetPoliciesDirPath: globalParams.FleetPoliciesDirPath,
-		}
-	},
-		"processchecks",
-		checkAllowlist,
-		getCoreAgentFxOptions,
-	)
-	return []*cobra.Command{cmd}
 }
